added rock playing and eating
